httpie_nsof: use local time for cache invalidation
Currently we used the exp in the token in order to decide whether the
token is going to be invalidated soon or not. However in case there is a
time diff between the client and the server which creates the token we
might invalidate too early or, even worse, too late. Therefore this
commit no longer uses the token's exp. Instead it calculates locally the
time and stores it in the local cache.